HIP-62 | Abhijeet/Raghavi | 1. Moves PrescriptionGenerationRequest to models
